requirements.txt
# ===============================
# Core Data Science Stack
# ===============================
numpy>=1.21,<2.0
pandas>=1.3,<2.0
scikit-learn>=1.0,<1.4
joblib==1.1.0

# ===============================
# Gradient Boosting Models
# ===============================
xgboost==1.5.2
lightgbm==3.3.2
catboost==1.0.4

# ===============================
# Hyperparameter Optimization
# ===============================
optuna==2.10.0

# ===============================
# MLOps & Orchestration
# ===============================
zenml==0.35.1
mlflow>=2.0,<2.10

# ===============================
# Data Version Control
# ===============================
dvc>=2.50,<3.0
dagshub>=0.3.14

# ===============================
# Visualization / App (Optional)
# ===============================
streamlit==1.8.1
matplotlib>=3.5,<3.9
seaborn>=0.11

# ===============================
# Utilities
# ===============================
pyyaml>=6.0.3,<6.1
python-dotenv>=0.21
markupsafe==1.1.1
python-box
ensure